new front

+++ src/main/java/ru/omgtu/ivt/controller/DogController.java
@@ -7,6 +7,8 @@ import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import ru.omgtu.ivt.docs.breeds.BreedCreateRequest;
 import ru.omgtu.ivt.docs.breeds.BreedsListResponse;
 import ru.omgtu.ivt.docs.dogs.DogsCreateRequest;
@@ -15,6 +17,8 @@ import ru.omgtu.ivt.docs.dogs.DogsListResponse;
 import ru.omgtu.ivt.model.dto.BreedDTO;
 import ru.omgtu.ivt.model.dto.DogDTO;
 import ru.omgtu.ivt.model.entity.Dog;
+import ru.omgtu.ivt.model.entity.User;
+import ru.omgtu.ivt.repository.DogRepository;
 import ru.omgtu.ivt.service.DogService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -22,6 +26,7 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/dogs")
@@ -29,6 +34,7 @@ import java.util.List;
 @Tag(name = "Собаки", description = "API для управления собаками")
 public class DogController {
     private final DogService dogService;
+    private final DogRepository dogRepository;
 
     @PostMapping
     @PreAuthorize("hasAnyRole('USER', 'MODERATOR', 'ADMIN')")
@@ -88,6 +94,15 @@ public class DogController {
             @Parameter(description = "ID собаки", example = "45")
             @PathVariable Long id
     ) {
+        Optional<Dog> dogSR = dogRepository.findById(id);
+        if(dogSR.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        Dog dog = dogSR.get();
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!dog.getOwner().getId().equals(user.getId())) {
+            return ResponseEntity.status(403).build();
+        }
         dogService.deleteDog(id);
         return ResponseEntity.noContent().build();
     }
